move suppress_auto_time
