+++ data/behaviours/time_stampable.py
@@ -1,3 +1,5 @@
+import contextlib
+
 from django.db import models
 
 
@@ -20,3 +22,27 @@ class TimeStampable(models.Model):
 
     creation_date = models.DateTimeField(auto_now_add=True)
     modification_date = models.DateTimeField(auto_now=True)
+
+
+@contextlib.contextmanager
+def suppress_autotime(model, fields):
+    """
+    Décorateur pour annuler temporairement le auto_now et auto_now_add de certains champs
+    Copié depuis https://stackoverflow.com/questions/7499767/temporarily-disable-auto-now-auto-now-add
+    """
+    _original_values = {}
+    for field in model._meta.local_fields:
+        if field.name in fields:
+            _original_values[field.name] = {
+                "auto_now": field.auto_now,
+                "auto_now_add": field.auto_now_add,
+            }
+            field.auto_now = False
+            field.auto_now_add = False
+    try:
+        yield
+    finally:
+        for field in model._meta.local_fields:
+            if field.name in fields:
+                field.auto_now = _original_values[field.name]["auto_now"]
+                field.auto_now_add = _original_values[field.name]["auto_now_add"]
